Move the printing code to it's own file.
